# crates/trapfalld/src/openapi.yaml
openapi: 3.0.3
info:
  title: TrapFall API
  description: |
    Lightweight self-hosted error capture engine. Sentry SDK compatible.

    ## Authentication
    Dashboard API uses session cookies (`trapfall_session`). Obtain via `POST /api/0/auth/login`.

    Ingest API uses DSN key auth via `X-Sentry-Auth` header or `Authorization: Bearer <key>`.
  version: 0.0.3
  contact:
    name: TrapFall
    url: https://github.com/codecoradev/trapfall
  license:
    name: Apache-2.0
servers:
  - url: http://localhost:3000
    description: Local development
...
paths:
...
  /api/0/setup:
    get:
      tags: [Setup]
      summary: Check if setup is needed
      operationId: getSetupStatus
      responses:
        "200":
          description: Setup status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SetupStatus"
    post:
      tags: [Setup]
      summary: Create admin account (first run only)
      operationId: createSetup
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SetupRequest"
      responses:
        "201":
          description: Admin created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SetupResponse"
        "403":
          description: Setup already completed

  # ── Auth ─────────────────────────────────────────────────────────
  /api/0/auth/login:
    post:
      tags: [Auth]
      summary: Login
      operationId: login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Login successful
          headers:
            Set-Cookie:
              schema:
                type: string
                example: trapfall_session=uuid; HttpOnly; SameSite=Strict
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
        "401":
          description: Invalid credentials
  /api/0/auth/logout:
    post:
      tags: [Auth]
      summary: Logout
      operationId: logout
      security:
        - cookieAuth: []
      responses:
        "200":
          description: Logged out
  /api/0/auth/me:
    get:
      tags: [Auth]
      summary: Get current user
      operationId: getMe
      security:
        - cookieAuth: []
      responses:
        "200":
          description: Current user info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserInfo"
        "401":
          description: Not authenticated
  /api/0/auth/change-password:
    post:
      tags: [Auth]
      summary: Change password
      operationId: changePassword
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChangePasswordRequest"
      responses:
        "200":
          description: Password changed
        "401":
          description: Current password incorrect

  # ── Projects ────────────────────────────────────────────────────
  /api/0/projects:
    get:
      tags: [Projects]
      summary: List all projects
      operationId: listProjects
      security:
        - cookieAuth: []
      responses:
        "200":
          description: Project list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Project"
        "401":
          description: Not authenticated
    post:
      tags: [Projects]
      summary: Create a project
      operationId: createProject
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProjectRequest"
      responses:
        "201":
          description: Project created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
        "409":
          description: Slug already exists
        "401":
          description: Not authenticated
  /api/0/projects/{slug}:
    get:
      tags: [Projects]
      summary: Get project by slug
      operationId: getProject
      security:
        - cookieAuth: []
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Project detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
        "404":
          description: Not found

  # ── Issues ──────────────────────────────────────────────────────
  /api/0/projects/{slug}/issues:
    get:
      tags: [Issues]
      summary: List issues for a project
      operationId: listIssues
      security:
        - cookieAuth: []
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: per_page
          in: query
          schema:
            type: integer
            default: 20
        - name: status
          in: query
          schema:
            type: string
            enum: [unresolved, resolved, ignored]
        - name: level
          in: query
          schema:
            type: string
            enum: [fatal, error, warning, info, debug]
      responses:
        "200":
          description: Issue list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IssueListResponse"
  /api/0/issues/{issueId}:
    get:
      tags: [Issues]
      summary: Get issue detail
      operationId: getIssue
      security:
        - cookieAuth: []
      parameters:
        - name: issueId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Issue detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Issue"
        "404":
          description: Not found
  /api/0/issues/{issueId}/status:
    post:
      tags: [Issues]
      summary: Set issue status
      operationId: setIssueStatus
      security:
        - cookieAuth: []
      parameters:
        - name: issueId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SetStatusRequest"
      responses:
        "200":
          description: Status updated

  # ── Events ──────────────────────────────────────────────────────
  /api/0/issues/{issueId}/events:
    get:
      tags: [Events]
      summary: List events for an issue
      operationId: listEvents
      security:
        - cookieAuth: []
      parameters:
        - name: issueId
          in: path
          required: true
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: per_page
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Event list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EventListResponse"

  # ── Search ──────────────────────────────────────────────────────
  /api/0/projects/{slug}/search:
    get:
      tags: [Search]
      summary: Search issues
      operationId: searchIssues
      security:
        - cookieAuth: []
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
        - name: q
          in: query
          required: true
          schema:
            type: string
          description: Search query (substring match on titles)
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: per_page
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Search results
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IssueListResponse"

  # ── Alert Rules ─────────────────────────────────────────────────
  /api/0/projects/{slug}/rules:
    get:
      tags: [Alert Rules]
      summary: List alert rules
      operationId: listAlertRules
      security:
        - cookieAuth: []
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Alert rules list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AlertRule"
    post:
      tags: [Alert Rules]
      summary: Create alert rule
      operationId: createAlertRule
      security:
        - cookieAuth: []
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAlertRuleRequest"
      responses:
        "200":
          description: Rule created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AlertRule"
  /api/0/rules/{ruleId}:
    get:
      tags: [Alert Rules]
      summary: Get alert rule
      operationId: getAlertRule
      security:
        - cookieAuth: []
      parameters:
        - name: ruleId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Alert rule detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AlertRule"
        "404":
          description: Not found
    delete:
      tags: [Alert Rules]
      summary: Delete alert rule
      operationId: deleteAlertRule
      security:
        - cookieAuth: []
      parameters:
        - name: ruleId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Rule deleted
  /api/0/rules/{ruleId}/toggle:
    post:
      tags: [Alert Rules]
      summary: Toggle alert rule on/off
      operationId: toggleAlertRule
      security:
        - cookieAuth: []
      parameters:
        - name: ruleId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Rule toggled

  # ── Ingest ──────────────────────────────────────────────────────
  /api/{projectId}/envelope/:
    post:
      tags: [Ingest]
      summary: Submit Sentry envelope
      description: |
        Sentry-compatible envelope endpoint. Authenticates via DSN key.

        Supports both `X-Sentry-Auth` header and `Authorization: Bearer <key>`.
      operationId: ingestEnvelope
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: string
          description: Project UUID
      requestBody:
        required: true
        content:
          application/x-sentry-envelope:
            schema:
              type: string
              format: binary
      responses:
        "200":
          description: Events accepted
        "400":
          description: Invalid envelope
        "401":
          description: Invalid DSN key
        "404":
          description: Project not found
        "429":
          description: Rate limited

  # ── WebSocket ───────────────────────────────────────────────────
  /api/0/ws:
    get:
      tags: [System]
      summary: WebSocket real-time updates
      description: |
        Upgrade to WebSocket. Requires session cookie auth.
        Sends `IssueCreated` and `IssueUpdated` messages as JSON.
      operationId: websocket
      security:
        - cookieAuth: []
      responses:
        "101":
          description: Switching Protocols (WebSocket upgrade)
        "401":
          description: Not authenticated
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: trapfall_session
    dsnKey:
      type: apiKey
      in: header
      name: X-Sentry-Auth
  schemas:
    SetupStatus:
      type: object
      properties:
        needs_setup:
          type: boolean
    SetupRequest:
      type: object
      required: [email, name, password]
      properties:
        email:
          type: string
          format: email
        name:
          type: string
        password:
          type: string
          format: password
    SetupResponse:
      type: object
      properties:
        user:
          $ref: "#/components/schemas/UserInfo"
        project_slug:
          type: string
        dsn:
          type: string
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
    LoginResponse:
      type: object
      properties:
        user:
          $ref: "#/components/schemas/UserInfo"
    UserInfo:
      type: object
      properties:
        id:
          type: string
        email:
          type: string
        name:
          type: string
        role:
          type: string
        created_at:
          type: string
    ChangePasswordRequest:
      type: object
      required: [current_password, new_password]
      properties:
        current_password:
          type: string
        new_password:
          type: string
    Project:
      type: object
      properties:
        id:
          type: string
        slug:
          type: string
        name:
          type: string
        dsn:
          type: string
          description: |
            Sentry-compatible DSN. In list responses the secret key is masked
            (e.g. `https://3167cebd...366b@host/id`). Full DSN is returned
            only on create / rotate / per-project GET.
        created_at:
          type: string
    CreateProjectRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
        slug:
          type: string
          description: Auto-generated from name if omitted
    Issue:
      type: object
      properties:
        id:
          type: string
        project_id:
          type: string
        fingerprint:
          type: string
        title:
          type: string
        culprit:
          type: string
        status:
          type: string
          enum: [unresolved, resolved, ignored]
        level:
          type: string
          enum: [fatal, error, warning, info, debug]
        count:
          type: integer
        user_count:
          type: integer
        first_seen:
          type: string
        last_seen:
          type: string
    SetStatusRequest:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [unresolved, resolved, ignored]
...
    IssueListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Issue"
        total:
          type: integer
        page:
          type: integer
        per_page:
          type: integer
...